Update .env.example
This commit modifies how we retrieve the MongoDB connection string from Azure Key Vault. Previously, the Key Vault URI (`kvUri`) was pointing directly to the specific secret inside the Azure Key Vault. 

The `kvUri` now points to the Azure Key Vault itself, adhering to standard practice for using the Azure Key Vault SDK. The secret name is then passed to the `GetSecret` method of the `SecretClient`, which allows us to retrieve the secret value. 

This change improves code clarity and flexibility in secret management.